v1.5.4 - 1.21.11-neoforge: settings reset button, terrain classifier height overflow, malformed-entry resilience, contour interpolation fix, hue bar, render-complete marker, private icon copies, volatile/teardown hardening, analysis button default on

+++ src/test/java/caeruleusTait/world/preview/backend/export/TerrainMapExporterFactsExportTest.java
@@ -92,7 +92,7 @@ class TerrainMapExporterFactsExportTest {
                 (blockX, blockZ) -> (short) 1,
                 new TerrainMapExporter.IdTableResolver(
                         new TerrainCategory[]{null, TerrainCategory.OCEAN},
-                        new byte[]{0, (byte) 50}));
+                        new short[]{0, 50}));
         exportAndCheck(facts, TerrainCategory.OCEAN.pixelColor());
     }
 
@@ -103,7 +103,7 @@ class TerrainMapExporterFactsExportTest {
                 (blockX, blockZ) -> null,
                 new TerrainMapExporter.IdTableResolver(
                         new TerrainCategory[]{null, TerrainCategory.OCEAN},
-                        new byte[]{0, (byte) 50}));
+                        new short[]{0, 50}));
         exportAndCheck(facts, TerrainCategory.PLAINS.pixelColor());
     }
 
@@ -114,7 +114,7 @@ class TerrainMapExporterFactsExportTest {
                 (blockX, blockZ) -> (short) 7,
                 new TerrainMapExporter.IdTableResolver(
                         new TerrainCategory[]{null, TerrainCategory.OCEAN},
-                        new byte[]{0, (byte) 50}));
+                        new short[]{0, 50}));
         exportAndCheck(facts, TerrainCategory.PLAINS.pixelColor());
     }
 }
